Redirected small types to 4B types, loads to adds.
This is the beginning of a history rewrite annotation.
This is a rewrite of a commit initially made in the main repo.
Original hash: abf5d6843e9fd5326fda0e3b67ac70556d8c7bfc.

+++ samples/trie.cu
@@ -51,7 +51,7 @@ struct trie {
         // the flag will protect against multiple pointer updates
         simt::std::atomic_flag flag = ATOMIC_FLAG_INIT;
     } next[26];
-    simt::std::atomic<int> count = ATOMIC_VAR_INIT(0);
+    simt::std::atomic<short> count = ATOMIC_VAR_INIT(0);
 };
 __host__ __device__
 int index_of(char c) {
